Agente/requirements.txt
Flask
Werkzeug
flask-cors
pytest
requests
pycryptodome
python-dotenv